+++ CHANGELOG.md
@@ -2,12 +2,24 @@
 
 ## DEV
 
+## v0.7.0
+
+#### Breaking changes
+
+- Indexer operation returns a `%Algoliax.Reponse{}` (#51)
+
+#### Bug fix
+
+- Ensure clean up on reindex fail
+
 ## v0.6.0 - 2021-06-10
 
 #### Updates
+
 - Replace deprecated `hmac` function with new :crypto API `mac` function to support OTP-24
 
 #### Breaking changes
+
 - Drop support for OTP 21. OTP 22+ is now required.
 
 ## v0.5.0 - 2021-03-04